refactor: use external color handling
