+++ apis/theme/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "main": "src/index.js",
   "devDependencies": {
-    "d3-color": "3.1.0",
     "extend": "3.0.2",
-    "node-event-emitter": "0.0.1"
+    "node-event-emitter": "0.0.1",
+    "qlik-chart-modules": "0.76.0"
   }
 }
