+++ apis/theme/src/contraster.js
@@ -1,6 +1,5 @@
 /* eslint no-cond-assign: 0 */
-import luminance from './luminance';
-import contrast from './contrast';
+import { getContrastRatio, getRelativeLuminance } from 'qlik-chart-modules';
 
 const MAX_SIZE = 1000;
 
@@ -8,7 +7,7 @@ export default function colorFn(colors = ['#333333', '#ffffff']) {
   let cache = {};
   let n = 0;
 
-  const luminances = colors.map(luminance);
+  const luminances = colors.map(getRelativeLuminance);
 
   return {
     getBestContrastColor(colorString) {
@@ -17,9 +16,9 @@ export default function colorFn(colors = ['#333333', '#ffffff']) {
           cache = {};
           n = 0;
         }
-        const L = luminance(colorString);
+        const L = getRelativeLuminance(colorString);
 
-        const contrasts = luminances.map((lum) => contrast(L, lum));
+        const contrasts = luminances.map((lum) => getContrastRatio(L, lum));
         const c = colors[contrasts.indexOf(Math.max(...contrasts))];
 
         cache[colorString] = c;
