--- apis/theme/src/contraster/contrast.js
@@ -1,4 +1,0 @@
-// https://www.w3.org/TR/UNDERSTANDING-WCAG20/visual-audio-contrast-contrast.html#contrast-ratiodef
-export default function contrast(L1, L2) {
-  return +((Math.max(L1, L2) + 0.05) / (Math.min(L1, L2) + 0.05)).toFixed(5);
-}
