--- apis/theme/src/utils/color/index.js
@@ -1,3 +1,0 @@
-import Color from './color';
-
-export default Color;
